// src/app/shared/interfaces/register.interface.ts
// Generated by https://quicktype.io

export interface RegisterUserPayload {
  first_name: string;
  last_name: string;
  email: string;
  company_name: string;
  password: string;
  gender: string;
  phone_number: string;
  date_of_birth: string;
}


export interface RegisterUserResponse {
  user: User;
  message: string;
}

export interface User {
  name: string;
  lastname: string;
  company_name: string;
  phonenumber: string;
  email: string;
  updated_at: string;
  created_at: string;
  id: number;
}